created service + moved leaderboard data there
